+++ app/src/main/java/com/vanpra/composematerialdialogdemos/demos/DateTimeDialog.kt
@@ -1,7 +1,12 @@
 package com.vanpra.composematerialdialogdemos.demos
 
+import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.graphics.Color
 import com.vanpra.composematerialdialogdemos.DialogAndShowButton
+import com.vanpra.composematerialdialogs.datetime.TimePickerDefaults
+import com.vanpra.composematerialdialogs.datetime.TimePickerColors
 import com.vanpra.composematerialdialogs.datetime.datepicker
 import com.vanpra.composematerialdialogs.datetime.datetimepicker
 import com.vanpra.composematerialdialogs.datetime.timepicker
@@ -13,7 +18,25 @@ import com.vanpra.composematerialdialogs.datetime.timepicker
 @Composable
 fun DateTimeDialogDemo() {
     DialogAndShowButton(buttonText = "Time Picker Dialog") {
-        timepicker()
+        val purple = remember { Color(0xFF3700B3) }
+
+        val colors: TimePickerColors = if (isSystemInDarkTheme()) {
+            TimePickerDefaults.colors(
+                activeBackgroundColor = purple.copy(0.3f),
+                activeTextColor = Color.White,
+                selectorColor = purple,
+                inactiveBackgroundColor = Color(0xFF292929),
+            )
+        } else  {
+            TimePickerDefaults.colors(
+                inactiveBackgroundColor = Color.LightGray,
+                activeBackgroundColor = purple.copy(0.1f),
+                activeTextColor = purple,
+                selectorColor = purple
+            )
+        }
+
+        timepicker(colors = colors)
     }
 
     DialogAndShowButton(buttonText = "Date Picker Dialog") {
@@ -21,7 +44,6 @@ fun DateTimeDialogDemo() {
     }
 
     DialogAndShowButton(buttonText = "Date and Time Picker Dialog") {
-        // Date time picker has a custom title as it needs to show a back icon
         datetimepicker()
     }
 }