+++ datetime/src/main/java/com/vanpra/composematerialdialogs/datetime/TimePicker.kt
@@ -27,6 +27,8 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.CornerSize
 import androidx.compose.material.MaterialTheme
+import androidx.compose.material.SliderDefaults
+import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
@@ -63,6 +65,7 @@ import kotlin.math.cos
 import kotlin.math.pow
 import kotlin.math.sin
 
+/* Offset of the clock line and selected circle */
 private data class SelectedOffset(
     val lineOffset: Offset = Offset.Zero,
     val selectedOffset: Offset = Offset.Zero
@@ -74,6 +77,7 @@ internal class SimpleLocalTime(hour: Int, minute: Int, isAM: Boolean) {
     var minute by mutableStateOf(minute)
     var isAM by mutableStateOf(isAM)
 
+    /* Converts a SimpleLocalTime object to a LocalTime object */
     fun toLocalTime(): LocalTime {
         val fullHour = if (isAM && hour == 12) {
             0
@@ -86,6 +90,7 @@ internal class SimpleLocalTime(hour: Int, minute: Int, isAM: Boolean) {
     }
 
     companion object {
+        /* Initalises a SimpleLocalTime object from a LocalTime object */
         fun fromLocalTime(time: LocalTime): SimpleLocalTime {
             val isAM = time.hour < 12
             val hour = if (isAM && time.hour == 0) {
@@ -100,14 +105,38 @@ internal class SimpleLocalTime(hour: Int, minute: Int, isAM: Boolean) {
     }
 }
 
+/**
+ * Represents the colors used by a [timepicker] and its parts in different states
+ *
+ * See [TimePickerDefaults.colors] for the default implementation
+ */
 interface TimePickerColors {
     val border: BorderStroke
 
+    /**
+     * Gets the background color dependant on if the item is active or not
+     *
+     * @param active true if the component/item is selected and false otherwise
+     * @return background color as a State
+     */
     @Composable
     fun backgroundColor(active: Boolean): State<Color>
 
+    /**
+     * Gets the text color dependant on if the item is active or not
+     *
+     * @param active true if the component/item is selected and false otherwise
+     * @return text color as a State
+     */
     @Composable
     fun textColor(active: Boolean): State<Color>
+
+
+    fun selectorColor(): Color
+    fun selectorTextColor(): Color
+
+    @Composable
+    fun periodBackgroundColor(active: Boolean): State<Color>
 }
 
 private class DefaultTimePickerColors(
@@ -115,12 +144,16 @@ private class DefaultTimePickerColors(
     private val inactiveBackgroundColor: Color,
     private val activeTextColor: Color,
     private val inactiveTextColor: Color,
+    private val inactivePeriodBackground: Color,
+    private val selectorColor: Color,
+    private val selectorTextColor: Color,
     borderColor: Color
 ) : TimePickerColors {
     override val border = BorderStroke(1.dp, borderColor)
 
     @Composable
     override fun backgroundColor(active: Boolean): State<Color> {
+        SliderDefaults
         return rememberUpdatedState(if (active) activeBackgroundColor else inactiveBackgroundColor)
     }
 
@@ -128,22 +161,57 @@ private class DefaultTimePickerColors(
     override fun textColor(active: Boolean): State<Color> {
         return rememberUpdatedState(if (active) activeTextColor else inactiveTextColor)
     }
+
+    override fun selectorColor(): Color {
+        return selectorColor
+    }
+
+    override fun selectorTextColor(): Color {
+        return selectorTextColor
+    }
+
+    @Composable
+    override fun periodBackgroundColor(active: Boolean): State<Color> {
+        return rememberUpdatedState(if (active) activeBackgroundColor else inactivePeriodBackground)
+    }
 }
 
-object DialogDefaults {
+/**
+* Object to hold default values used by [timepicker]
+*/
+object TimePickerDefaults {
+    /**
+     * Initialises a [TimePickerColors] object which represents the different colors used by
+     * the [timepicker] composable
+     *
+     * @param activeBackgroundColor background color of selected time unit or period (AM/PM)
+     * @param inactiveBackgroundColor background color of inactive items in the dialog including the clock face
+     * @param activeTextColor color of text on the activeBackgroundColor
+     * @param inactiveTextColor color of text on the inactiveBackgroundColor
+     * @param inactivePeriodBackground background color of the inactive period (AM/PM) selector
+     * @param borderColor border color of the period (AM/PM) selector
+     * @param selectorColor color of clock hand/selector
+     * @param selectorTextColor color of text on selectedColor
+     */
     @Composable
-    fun timePickerColors(
+    fun colors(
         activeBackgroundColor: Color = MaterialTheme.colors.primary.copy(0.3f),
         inactiveBackgroundColor: Color = MaterialTheme.colors.onBackground.copy(0.3f),
         activeTextColor: Color = MaterialTheme.colors.onPrimary,
         inactiveTextColor: Color = MaterialTheme.colors.onBackground,
-        borderColor: Color = MaterialTheme.colors.onBackground
+        inactivePeriodBackground: Color = Color.Transparent,
+        borderColor: Color = MaterialTheme.colors.onBackground,
+        selectorColor: Color = MaterialTheme.colors.primary,
+        selectorTextColor: Color = MaterialTheme.colors.onPrimary
     ): TimePickerColors {
         return DefaultTimePickerColors(
             activeBackgroundColor = activeBackgroundColor,
             inactiveBackgroundColor = inactiveBackgroundColor,
             activeTextColor = activeTextColor,
             inactiveTextColor = inactiveTextColor,
+            inactivePeriodBackground = inactivePeriodBackground,
+            selectorColor = selectorColor,
+            selectorTextColor = selectorTextColor,
             borderColor = borderColor
         )
     }
@@ -187,7 +255,7 @@ internal class TimePickerState(
 @Composable
 fun MaterialDialog.timepicker(
     initialTime: LocalTime = LocalTime.now(),
-    colors: TimePickerColors = DialogDefaults.timePickerColors(),
+    colors: TimePickerColors = TimePickerDefaults.colors(),
     onCancel: () -> Unit = {},
     onComplete: (LocalTime) -> Unit = {}
 ) {
@@ -222,7 +290,8 @@ internal fun TimePickerImpl(
                     label = { index -> if (index == 0) "12" else index.toString() },
                     onAnchorChange = { hours -> state.selectedTime.hour = hours },
                     startAnchor = state.selectedTime.hour,
-                    onLift = { state.currentScreen = ClockScreen.Minute }
+                    onLift = { state.currentScreen = ClockScreen.Minute },
+                    colors = state.colors
                 )
 
                 ClockScreen.Minute -> ClockLayout(
@@ -230,7 +299,8 @@ internal fun TimePickerImpl(
                     label = { index -> index.toString().padStart(2, '0') },
                     onAnchorChange = { mins -> state.selectedTime.minute = mins },
                     startAnchor = state.selectedTime.minute,
-                    isNamedAnchor = { anchor -> anchor % 5 == 0 }
+                    isNamedAnchor = { anchor -> anchor % 5 == 0 },
+                    colors = state.colors
                 )
             }
         }
@@ -263,7 +333,7 @@ internal fun TimePickerTitle(onBack: (() -> Unit)?) {
                 style = TextStyle(color = MaterialTheme.colors.onBackground)
             )
         }
-    } else  {
+    } else {
         Box(Modifier.height(52.dp)) {
             Text(
                 "SELECT TIME",
@@ -282,19 +352,21 @@ internal fun ClockLabel(
     textColor: Color,
     onClick: () -> Unit
 ) {
-    Box(
-        Modifier.width(96.dp).fillMaxHeight().clip(MaterialTheme.shapes.medium)
-            .background(backgroundColor)
-            .clickable(onClick = onClick),
-        contentAlignment = Alignment.Center
+
+    Surface(
+        modifier = Modifier.width(96.dp).fillMaxHeight().clickable(onClick = onClick),
+        shape = MaterialTheme.shapes.medium,
+        color = backgroundColor,
     ) {
-        Text(
-            text = text,
-            style = TextStyle(
-                fontSize = 50.sp,
-                color = textColor
+        Box(modifier = Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+            Text(
+                text = text,
+                style = TextStyle(
+                    fontSize = 50.sp,
+                    color = textColor
+                )
             )
-        )
+        }
     }
 }
 
@@ -340,7 +412,7 @@ internal fun TimeLayout(state: TimePickerState) {
             Box(
                 modifier = Modifier.size(height = 40.dp, width = 52.dp)
                     .clip(topPeriodShape)
-                    .background(state.colors.backgroundColor(state.selectedTime.isAM).value)
+                    .background(state.colors.periodBackgroundColor(state.selectedTime.isAM).value)
                     .clickable { state.selectedTime.isAM = true },
                 contentAlignment = Alignment.Center
             ) {
@@ -352,7 +424,7 @@ internal fun TimeLayout(state: TimePickerState) {
             Box(
                 modifier = Modifier.size(height = 40.dp, width = 52.dp)
                     .clip(bottomPeriodShape)
-                    .background(state.colors.backgroundColor(!state.selectedTime.isAM).value)
+                    .background(state.colors.periodBackgroundColor(!state.selectedTime.isAM).value)
                     .clickable { state.selectedTime.isAM = false },
                 contentAlignment = Alignment.Center
             ) {
@@ -371,6 +443,7 @@ private fun ClockLayout(
     anchorPoints: Int,
     label: (Int) -> String,
     startAnchor: Int,
+    colors: TimePickerColors,
     onAnchorChange: (Int) -> Unit = {},
     onLift: () -> Unit = {}
 ) {
@@ -380,6 +453,7 @@ private fun ClockLayout(
     val offset = remember { mutableStateOf(Offset.Zero) }
     val center = remember { mutableStateOf(Offset.Zero) }
     val namedAnchor = remember { mutableStateOf(isNamedAnchor(startAnchor)) }
+    val selectedAnchor = remember { mutableStateOf(startAnchor) }
 
     val anchors = remember {
         val anchors = mutableListOf<SelectedOffset>()
@@ -413,6 +487,7 @@ private fun ClockLayout(
 
             anchoredOffset.value = anchors[minAnchor]
             namedAnchor.value = isNamedAnchor(minAnchor)
+            selectedAnchor.value = minAnchor
         }
     }
 
@@ -447,16 +522,17 @@ private fun ClockLayout(
             offset.value = center.value
         }
 
-        val textColor = MaterialTheme.colors.onBackground.toArgb()
-        val selectedColor = MaterialTheme.colors.primary
-        val clockSurfaceColor = MaterialTheme.colors.onBackground.copy(0.3f)
-        val clockSurfaceDiameter = constraints.maxWidth.toFloat() / 2f
+        val inactiveTextColor = colors.textColor(false).value.toArgb()
+        val clockBackgroundColor = colors.backgroundColor(false).value
+        val selectorColor = remember { colors.selectorColor() }
+        val selectorTextColor = remember { colors.selectorTextColor().toArgb() }
+        val clockSurfaceDiameter = remember(constraints.maxWidth) { constraints.maxWidth.toFloat() / 2f }
 
         Canvas(modifier = Modifier.fillMaxSize()) {
-            drawCircle(clockSurfaceColor, radius = clockSurfaceDiameter, center = center.value)
-            drawCircle(selectedColor, radius = 16f, center = center.value)
+            drawCircle(clockBackgroundColor, radius = clockSurfaceDiameter, center = center.value)
+            drawCircle(selectorColor, radius = 16f, center = center.value)
             drawLine(
-                color = selectedColor,
+                color = selectorColor,
                 start = center.value,
                 end = center.value + anchoredOffset.value.lineOffset,
                 strokeWidth = 10f,
@@ -464,7 +540,7 @@ private fun ClockLayout(
             )
 
             drawCircle(
-                selectedColor,
+                selectorColor,
                 center = center.value + anchoredOffset.value.selectedOffset,
                 radius = selectedRadius,
                 alpha = 0.7f
@@ -483,6 +559,11 @@ private fun ClockLayout(
                 for (x in 0 until 12) {
                     val angle = (2 * PI / 12) * (x - 15)
                     val textOuter = label(x * anchorPoints / 12)
+                    val textColor = if (selectedAnchor.value == textOuter.toInt()) {
+                        selectorTextColor
+                    } else {
+                        inactiveTextColor
+                    }
 
                     drawText(
                         60f,